fix,refactor: consolidate zebrafish code into build script
changelog:
- refactor: previously, zebrafish chemical and sample file generation code was occurring across two separate files (map_samples_to_chemicals.py and build_script.py). Sample files were being produced by the build_script.py, but were not being produced correctly. Code has been updated to remove the code generating incorrect zebrafish sample files and moving all zebrafish file combination code to build_script.py. This also involved moving code dependencies to src/samples.py.
- fix: fix the generation of incorrect sample files (see above)
